Сделал 2 часть ДЗ 8 модуля
