+++ js/big-picture.js
@@ -1,21 +1,35 @@
+import { showComments } from './comments.js';
 import { isEscapeKey } from './utils.js';
 
-const MAX_COMMENTS_SHOWN = 2;
 const bigPictureElement = document.querySelector('.big-picture');
 const closeBigPicture = bigPictureElement.querySelector('#picture-cancel');
+const commentsLoaderElement = bigPictureElement.querySelector('.social__comments-loader');
+
+let comments = [];
+
+const onLoaderClick = (evt) => {
+  evt.preventDefault();
+  showComments(comments);
+};
+commentsLoaderElement.addEventListener('click', onLoaderClick);
+
+const hideBigPicture = () => {
+  bigPictureElement.classList.add('hidden');
+  document.body.classList.remove('modal-open');
+  document.removeEventListener('keydown', onDocumentKeydown);
+};
 
 const onPictureCancelButtonClick = () => {
   hideBigPicture();
 };
-
 closeBigPicture.addEventListener('click', onPictureCancelButtonClick);
 
-const onDocumentKeydown = (evt) => {
+function onDocumentKeydown(evt) {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     hideBigPicture();
   }
-};
+}
 
 const showBigPicture = (target) => {
   const picture = bigPictureElement.querySelector('.big-picture__img img');
@@ -25,38 +39,12 @@ const showBigPicture = (target) => {
   bigPictureElement.querySelector('.social__caption').textContent = target.description;
   bigPictureElement.querySelector('.likes-count').textContent = target.likes;
 
-  const comments = target.comments.length;
-  const commentsShown = comments > MAX_COMMENTS_SHOWN ? MAX_COMMENTS_SHOWN : comments;
-  bigPictureElement.querySelector('.social__comment-total-count').textContent = comments;
-  bigPictureElement.querySelector('.social__comment-shown-count').textContent = commentsShown;
+  comments = target.comments;
+  showComments(comments);
 
-  const commentsContainer = bigPictureElement.querySelector('.social__comments');
-  commentsContainer.innerHTML = '';
-  const commentTemplate = document.querySelector('#comment')
-    .content
-    .querySelector('.social__comment');
-  const commentsFragment = document.createDocumentFragment();
-
-  for(let i = 0;i < commentsShown;i++) {
-    const commentElement = commentTemplate.cloneNode(true);
-    commentElement.querySelector('.social__picture').src = target.comments[i].avatar;
-    commentElement.querySelector('.social__picture').alt = target.comments[i].name;
-    commentElement.querySelector('.social__text').textContent = target.comments[i].message;
-    commentsFragment.append(commentElement);
-  }
-  commentsContainer.append(commentsFragment);
-
-  bigPictureElement.querySelector('.social__comment-count').classList.add('hidden');
-  bigPictureElement.querySelector('.comments-loader').classList.add('hidden');
   bigPictureElement.classList.remove('hidden');
   document.body.classList.add('modal-open');
   document.addEventListener('keydown', onDocumentKeydown);
 };
-
-function hideBigPicture() {
-  bigPictureElement.classList.add('hidden');
-  document.body.classList.remove('modal-open');
-  document.removeEventListener('keydown', onDocumentKeydown);
-}
 
 export {showBigPicture};
