+++ package.json
@@ -17,7 +17,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "start": "vite --port 3000 --open",
+    "start": "vite --port 3000",
     "lint": "eslint js/"
   },
   "engines": {
